[Feature] Better sum types (#4)
* [Feature] Better sum types

Throwing away the previous implementation of sum types, which didn't quite work.

* [Fix][Test] Consolidating equality checks

Our tests were flaky because we rely upon JSON equality but in practice,
in Object or OneOf, the order or elements will change across runs. Implementing
manual equality checm for the elements of schema.*, which should fix the
flakiness.